Add files via upload
This is the v2 (second version) about ramanujan's number code. Now, we've improved the CSS file to let him looks attractive. And also, it's more dynamic than the previous version (v1: essay1.html); so now, this new version help user to see the operation who justify his the validity of his number (if it's truth).