Update essay2.html
This is the v2 of our code about Ramanujan's number. In this new version, the website is more interactive and dynamic. In this new version, we've update the CSS part to make the page more attractive than the previous; about the JAVASCRIPT part, the code can now show you the operation to validate your truth number. The page looks like better than the v1.